fix: float32 casting for imudata

+++ src/models/sensors.py
@@ -17,8 +17,12 @@ class EscData(BaseModel):
 class ImuData(BaseModel):
     """Model for IMU data."""
 
-    acceleration: NDArray[Shape["3"], np.float32] = np.array([0.0, 0.0, 0.0])  # pyright: ignore[reportGeneralTypeIssues]
-    gyroscope: NDArray[Shape["3"], np.float32] = np.array([0.0, 0.0, 0.0])  # pyright: ignore[reportGeneralTypeIssues]
+    acceleration: NDArray[Shape["3"], np.float32] = np.array(  # pyright: ignore[reportGeneralTypeIssues]
+        [0.0, 0.0, 0.0], dtype=np.float32
+    )
+    gyroscope: NDArray[Shape["3"], np.float32] = np.array(  # pyright: ignore[reportGeneralTypeIssues]
+        [0.0, 0.0, 0.0], dtype=np.float32
+    )
     temperature: float = 0.0
