+++ src/sensors/imu.py
@@ -21,6 +21,7 @@ from bmi270.BMI270 import (
     I2C_PRIM_ADDR,
     PERFORMANCE_MODE,
 )
+import numpy as np
 
 from ..constants import SYSTEM_FAILURE_THRESHOLD
 from ..log import log_error, log_info
@@ -80,8 +81,8 @@ class Imu:
             return None
         try:
             return ImuData(
-                acceleration=self.imu.get_acc_data(),
-                gyroscope=self.imu.get_gyr_data(),
+                acceleration=self.imu.get_acc_data().astype(np.float32),
+                gyroscope=self.imu.get_gyr_data().astype(np.float32),
                 temperature=self.imu.get_temp_data(),
             )
         except Exception as e:
